Refactor application structure by creating HomePage, GamePage, and ScorePage components
